chore: removes usage of memory package in integration and system tests
